# NameSpaceHelper: introduce namespace cache to speedup namespace lookups