This pr introduce cache for namespace lookups.

This speed up for 8kloc test-case
- ReferenceThrowableOnlySniff from 3.8s to 0.4s
- OptimizedFunctionsWithoutUnpackingSniff from 3.5s to 0.4s
- ParameterTypeHintSniff from 2.8s to 0.4s
- ReturnTypeHintSniff from 1.7s to 0.4s

Can this cache cause any unwanted side effects?