@kukulich 2020-03-25
@nightlinus One last change please. Squash it to one commit or move this change https://github.com/slevomat/coding-standard/pull/939/commits/3d0441a85952c3b9bd36460bdd49472d50c98290#diff-054ee9cafd168683808f20c4b5d722f1R84 to the right commit.